refactored to entirely separate component
