+++ src/index.ts
@@ -67,7 +67,7 @@ export { RangeInput } from './components/forms/RangeInput/RangeInput'
 export { Select } from './components/forms/Select/Select'
 export { Textarea } from './components/forms/Textarea/Textarea'
 export { TextInput } from './components/forms/TextInput/TextInput'
-export { TextInputMask } from './components/forms/TextInput/TextInputMask'
+export { TextInputMask } from './components/forms/TextInputMask/TextInputMask'
 export { TimePicker } from './components/forms/TimePicker/TimePicker'
 export { ValidationChecklist } from './components/forms/Validation/ValidationChecklist'
 export { ValidationItem } from './components/forms/Validation/ValidationItem'
